% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curatedAdipRNA.R
\docType{package}
\name{curatedAdipoRNA}
\alias{curatedAdipoRNA}
\alias{curatedAdipoRNA-package}
\title{\code{curatedAdipoRNA} package}
\description{
A curated RNA-Seq Dataset of MDI-induced Differentiated Adipocytes (3T3-L1)
}
\details{
A curated dataset of publicly available RNA-sequencing of the 3T3-L1
pre-adipocyte cell line. The package document the data collection,
pre-processing and processing of the data.
}
\section{Processed data and metadata}{

\code{samples.csv}
\code{runs.csv}
\code{studies.bib}
\code{gene_body.txt}
}

\section{Pre-processing and processing scripts}{
NA
}

\section{Counts and normalized counts object \code{counts.rda}}{
NA
}

