% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_counts}
\alias{gene_counts}
\title{Gene counts of differentiating adipocytes}
\format{\code{\link[SummarizedExperiment]{SummarizedExperiment}}}
\usage{
gene_counts
}
\description{
Gene counts of differentiating adipocytes
}
\keyword{datasets}
