% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_counts}
\alias{gene_counts}
\title{Gene counts of differentiating adipocytes}
\format{A \code{RangedSummarizedExperiment} object
contains:

\describe{
\item{assay}{The gene counts \code{matrix}.}
\item{colData}{The phenotype data.}
\item{rowRanges}{The feature data}
\item{metadata}{The metadata \code{metadata} which contain two objects;
\code{qc} A \code{list} of \code{\link[fastqcr]{qc_read}} objects of
each \code{fastq} file and a \code{data.frame} of the studies from which
the samples were collected.}
}}
\usage{
gene_counts
}
\description{
Gene counts of differentiating adipocytes
}
\examples{
# load the data object
data('gene_counts')

# print the object
gene_counts
}
\keyword{datasets}
